python-pipeline/requirements.txt
pandas
numpy
requests
psycopg2-binary
sqlalchemy
python-dotenv
pydantic-settings
pydantic
google-generativeai
rapidfuzz
python-dateutil
coloredlogs
